# frontend/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Diverge API",
    "version": "1.0.0",
    "description": "Frozen HTTP API contract for Diverge financial narrative intelligence, sentiment tracking, multi-index calculation, integrity scoring, and reasoning trace explainability. This contract is frozen between backend (Antigravity) and frontend (Cursor). Any modifications must be called out explicitly."
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/ticker/{symbol}": {
      "get": {
        "tags": ["Tickers & Detail"],
        "summary": "Full Advanced-mode ticker details",
        "description": "Returns comprehensive diagnostic view for a ticker including all 5 financial indices (Contagion Core, Reality Check, Capitulation Signal, Cross-Market Spillover, Language Divergence), Trust Score, composite metrics, and narrative context. Every index strictly adheres to the { score, available, coverage_note } pattern with approved human-readable labels.",
        "operationId": "getTickerDetail",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "description": "Standard uppercase ticker symbol (e.g., INFY, AAPL, RELIANCE)",
            "schema": {
              "type": "string",
              "example": "INFY"
            }
          },
          {
            "name": "window",
            "in": "query",
            "required": false,
            "description": "Optional ISO 8601 UTC timestamp of window start. Defaults to latest window.",
            "schema": {
              "type": "string",
              "format": "date-time",
              "example": "2026-06-11T19:40:24.121335Z"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Full Advanced-mode technical metrics",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TickerAdvancedResponse"
                }
              }
            }
          },
          "404": {
            "description": "Ticker or requested metric window not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/ticker/{symbol}/simple": {
      "get": {
        "tags": ["Tickers & Detail"],
        "summary": "Simple-mode plain-language summary",
        "description": "Returns non-technical, plain-language consumer summary for a ticker. Excludes raw index numbers; provides qualitative verdict, composite score, explanatory synthesis sentence, and confidence rating.",
        "operationId": "getTickerSimple",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "description": "Standard uppercase ticker symbol (e.g., INFY, AAPL, RELIANCE)",
            "schema": {
              "type": "string",
              "example": "INFY"
            }
          },
          {
            "name": "window",
            "in": "query",
            "required": false,
            "description": "Optional ISO 8601 UTC timestamp of window start. Defaults to latest window.",
            "schema": {
              "type": "string",
              "format": "date-time",
              "example": "2026-06-11T19:40:24.121335Z"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Simple-mode plain-language summary",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TickerSimpleResponse"
                }
              }
            }
          },
          "404": {
            "description": "Ticker or requested metric window not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/ticker/{symbol}/reasoning": {
      "get": {
        "tags": ["Explainability"],
        "summary": "Post-level reasoning trace audit trail",
        "description": "Returns the granular post-level explainability trace ('Why this score?') for a ticker window. Groups source evidence posts into categorized audit drivers (virality onset, cross-market sentiment, regional vernacular, capitulation signals, near-duplicates, uncharacteristic consensus).",
        "operationId": "getTickerReasoningTrace",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "description": "Standard uppercase ticker symbol (e.g., INFY, AAPL)",
            "schema": {
              "type": "string",
              "example": "INFY"
            }
          },
          {
            "name": "window",
            "in": "query",
            "required": false,
            "description": "Optional ISO 8601 UTC timestamp of window start. Defaults to latest window.",
            "schema": {
              "type": "string",
              "format": "date-time",
              "example": "2026-06-11T19:40:24.121335Z"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reasoning trace audit trail panel data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReasoningTraceResponse"
                }
              }
            }
          },
          "404": {
            "description": "Ticker or reasoning traces not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Standard Bearer JWT Authorization header (e.g. 'Bearer <token>')"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "required": ["error_code", "message"],
        "properties": {
          "error_code": {
            "type": "string",
            "example": "INVALID_CREDENTIALS"
          },
          "message": {
            "type": "string",
            "example": "Invalid email or password provided."
          },
          "details": {
            "type": "object",
            "nullable": true,
            "additionalProperties": true,
            "example": {
              "field": "email"
            }
          }
        }
      },
...
      "IndexMetricValue": {
        "type": "object",
        "required": ["label", "score", "available", "coverage_note"],
        "description": "Standardized index metric envelope. If available is false, score is null and coverage_note explains the data requirement.",
        "properties": {
          "label": {
            "type": "string",
            "description": "Approved human-readable name for the index metric",
            "example": "Contagion Core"
          },
          "score": {
            "type": "number",
            "format": "float",
            "nullable": true,
            "description": "Numeric computed score, or null when unavailable",
            "example": 1.34
          },
          "available": {
            "type": "boolean",
            "description": "True if sufficient valid data was available to compute the index",
            "example": true
          },
          "coverage_note": {
            "type": "string",
            "nullable": true,
            "description": "Explicit reason if the index could not be computed, otherwise null",
            "example": null
          }
        }
      },
      "TickerSimpleItem": {
        "type": "object",
        "required": ["id", "symbol", "name", "sector", "score", "verdict_label", "why_sentence", "trust_label", "window_start_utc"],
        "properties": {
          "id": {
            "type": "string",
            "example": "INFY"
          },
          "symbol": {
            "type": "string",
            "example": "INFY"
          },
          "name": {
            "type": "string",
            "example": "Infosys Ltd."
          },
          "sector": {
            "type": "string",
            "example": "Information Technology"
          },
          "score": {
            "type": "number",
            "format": "float",
            "nullable": true,
            "minimum": 0,
            "maximum": 100,
            "description": "Calibrated composite confidence score (0-100), or null if insufficient data",
            "example": 54.6
          },
          "verdict_label": {
            "type": "string",
            "enum": ["building", "mixed", "fading", "insufficient_data"],
            "example": "mixed"
          },
          "why_sentence": {
            "type": "string",
            "description": "Synthesized plain-language summary of what is driving the sentiment score",
            "example": "Narrative momentum driven by virality rate (Rn 1.34) — consumer sentiment indicates market may be underpricing reality."
          },
          "trust_label": {
            "type": "string",
            "example": "High confidence"
          },
          "window_start_utc": {
            "type": "string",
            "format": "date-time",
            "example": "2026-06-11T19:40:24.121335Z"
          }
        }
      },
...
      "TickerSimpleResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/TickerSimpleItem"
          }
        ]
      },
      "IndicesBreakdown": {
        "type": "object",
        "required": ["rn", "cirg", "cli", "cassi", "vdi"],
        "description": "All 5 financial indices adhering to the frozen null-handling and approved human-readable label rules.",
        "properties": {
          "rn": {
            "allOf": [{ "$ref": "#/components/schemas/IndexMetricValue" }],
            "description": "Contagion Core (Rn) — Epidemiological viral narrative reproduction rate"
          },
          "cirg": {
            "allOf": [{ "$ref": "#/components/schemas/IndexMetricValue" }],
            "description": "Reality Check (CIRG) — Consumer app/product reviews vs. investor sentiment gap"
          },
          "cli": {
            "allOf": [{ "$ref": "#/components/schemas/IndexMetricValue" }],
            "description": "Capitulation Signal (CLI) — Despair, liquidation, and stop-loss panic density"
          },
          "cassi": {
            "allOf": [{ "$ref": "#/components/schemas/IndexMetricValue" }],
            "description": "Cross-Market Spillover (CASSI) — VAR cross-asset sentiment transmission"
          },
          "vdi": {
            "allOf": [{ "$ref": "#/components/schemas/IndexMetricValue" }],
            "description": "Language Divergence (VDI) — Regional vernacular vs. institutional English sentiment"
          }
        }
      },
      "CoordinationMetric": {
        "type": "object",
        "required": ["label", "coordination_score", "confidence_flag", "available", "coverage_note"],
        "description": "Inorganic coordination and trust score evaluation.",
        "properties": {
          "label": {
            "type": "string",
            "example": "Trust Score"
          },
          "coordination_score": {
            "type": "number",
            "format": "float",
            "minimum": 0,
            "maximum": 100,
            "description": "0-100 inorganic coordination suspicion score. Higher indicates bot/coordinated behavior.",
            "example": 32.5
          },
          "confidence_flag": {
            "type": "string",
            "enum": ["high_trust", "moderate", "low_trust", "insufficient_data"],
            "example": "high_trust"
          },
          "available": {
            "type": "boolean",
            "example": true
          },
          "coverage_note": {
            "type": "string",
            "nullable": true,
            "example": null
          }
        }
      },
      "PhylogenyTransitionItem": {
        "type": "object",
        "required": ["window_start_utc", "mutation_type", "mutation_detail", "composite_delta"],
        "properties": {
          "window_start_utc": {
            "type": "string",
            "format": "date-time",
            "example": "2026-06-11T19:40:24.121335Z"
          },
          "mutation_type": {
            "type": "string",
            "enum": ["stable", "dominant_index_shift", "composite_reversal", "new_risk_flag", "flag_resolved"],
            "example": "stable"
          },
          "mutation_detail": {
            "type": "object",
            "additionalProperties": true,
            "example": {
              "current_dominant": "rn",
              "prior_dominant": "rn"
            }
          },
          "composite_delta": {
            "type": "number",
            "format": "float",
            "nullable": true,
            "example": 1.1
          }
        }
      },
      "TickerAdvancedResponse": {
        "type": "object",
        "required": [
          "symbol",
          "name",
          "sector",
          "window_start_utc",
          "window_end_utc",
          "composite_score",
          "dominant_index",
          "risk_flags",
          "aggregation_confidence",
          "indices",
          "coordination",
          "phylogeny_context"
        ],
        "properties": {
          "symbol": {
            "type": "string",
            "example": "INFY"
          },
          "name": {
            "type": "string",
            "example": "Infosys Ltd."
          },
          "sector": {
            "type": "string",
            "example": "Information Technology"
          },
          "window_start_utc": {
            "type": "string",
            "format": "date-time",
            "example": "2026-06-11T19:40:24.121335Z"
          },
          "window_end_utc": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "example": "2026-09-10T01:14:12.000000Z"
          },
          "composite_score": {
            "type": "number",
            "format": "float",
            "minimum": 0,
            "maximum": 100,
            "nullable": true,
            "example": 54.6
          },
          "dominant_index": {
            "type": "string",
            "example": "rn"
          },
          "risk_flags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["consumer_reality_underpriced"]
          },
          "aggregation_confidence": {
            "type": "string",
            "enum": ["high_trust", "moderate", "low_trust", "insufficient_data"],
            "example": "high_trust"
          },
          "indices": {
            "$ref": "#/components/schemas/IndicesBreakdown"
          },
          "coordination": {
            "$ref": "#/components/schemas/CoordinationMetric"
          },
          "phylogeny_context": {
            "type": "array",
            "description": "Recent narrative lineage transitions leading up to this window",
            "items": {
              "$ref": "#/components/schemas/PhylogenyTransitionItem"
            }
          }
        }
      },
      "ReasoningTraceItem": {
        "type": "object",
        "required": ["trace_id", "post_id", "account_id", "platform", "weight", "upvotes", "text_preview"],
        "properties": {
          "trace_id": {
            "type": "integer",
            "example": 1042
          },
          "post_id": {
            "type": "string",
            "example": "tw_182937461928"
          },
          "account_id": {
            "type": "string",
            "example": "trader_daily_in"
          },
          "platform": {
            "type": "string",
            "enum": ["stocktwits", "reddit", "telegram", "rss_news", "google_trends"],
            "example": "stocktwits"
          },
          "weight": {
            "type": "number",
            "format": "float",
            "example": 0.85
          },
          "upvotes": {
            "type": "integer",
            "example": 14
          },
          "text_preview": {
            "type": "string",
            "maxLength": 145,
            "example": "INFY Q2 margins expanding faster than expected, digital cloud pipeline strong into next quarter..."
          }
        }
      },
      "ReasoningTraceResponse": {
        "type": "object",
        "required": ["symbol", "window_start_utc", "total_traces", "categories"],
        "properties": {
          "symbol": {
            "type": "string",
            "example": "INFY"
          },
          "window_start_utc": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "example": "2026-06-11T19:40:24.121335Z"
          },
          "total_traces": {
            "type": "integer",
            "example": 28
          },
          "categories": {
            "type": "object",
            "description": "Audit driver categories containing source evidence posts",
            "properties": {
              "rn_onset": {
                "type": "array",
                "items": { "$ref": "#/components/schemas/ReasoningTraceItem" }
              },
              "cassi_sentiment": {
                "type": "array",
                "items": { "$ref": "#/components/schemas/ReasoningTraceItem" }
              },
              "vdi_divergence": {
                "type": "array",
                "items": { "$ref": "#/components/schemas/ReasoningTraceItem" }
              },
              "cli_capitulation": {
                "type": "array",
                "items": { "$ref": "#/components/schemas/ReasoningTraceItem" }
              },
              "duplicate_flag": {
                "type": "array",
                "items": { "$ref": "#/components/schemas/ReasoningTraceItem" }
              },
              "sentiment_variance_outlier": {
                "type": "array",
                "items": { "$ref": "#/components/schemas/ReasoningTraceItem" }
              }
            },
            "additionalProperties": {
              "type": "array",
              "items": { "$ref": "#/components/schemas/ReasoningTraceItem" }
            }
          }
        }
      }
    }
  }
}